+++ src/Controller/UsersController.php
@@ -2,12 +2,15 @@
 
 namespace App\Controller;
 
+use App\Controller\Traits\AdminAccessTrait;
 use App\Entity\Eleve;
 use App\Entity\Utilisateur;
 use App\Form\EleveType;
 use App\Form\UtilisateurType;
 use App\Repository\EleveRepository;
+use App\Repository\StageRepository;
 use App\Repository\UtilisateurRepository;
+use Doctrine\DBAL\Exception\ForeignKeyConstraintViolationException;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -16,34 +19,34 @@ use Symfony\Component\Routing\Attribute\Route;
 
 final class UsersController extends AbstractController
 {
+    use AdminAccessTrait;
+
     #[Route('/utilisateurs', name: 'app_users')]
     #[Route(name: 'app_eleve_index', methods: ['GET'])]
     public function indexEleve(EleveRepository $eleveRepository, UtilisateurRepository $utilisateurRepository): Response
     {
-        if (!$this->isGranted('ROLE_ADMIN')) {
-            return $this->redirectToRoute('app_access_denied');
+        if ($response = $this->redirectIfNotAdmin()) {
+            return $response;
         }
 
-        return $this->render('eleve/index.html.twig', [
+        return $this->render('home/utilisateur.html.twig', [
             'eleves' => $eleveRepository->findAll(),
             'utilisateurs' => $utilisateurRepository->findAll(),
         ]);
     }
 
-    // CRUD pour les élèves
-    // Création d'un nouvel élève
     #[Route('/newEleve', name: 'app_eleve_new', methods: ['GET', 'POST'])]
     public function newEleve(Request $request, EntityManagerInterface $entityManager): Response
     {
         $eleve = new Eleve();
-        $form = $this->createForm(EleveType::class, $eleve);
+        $form = $this->createEleveForm($eleve);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->persist($eleve);
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_users', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToUsersIndex();
         }
 
         return $this->render('eleve/new.html.twig', [
@@ -52,7 +55,6 @@ final class UsersController extends AbstractController
         ]);
     }
 
-    // Affichage des détails d'un élève
     #[Route('/Eleve/{id}', name: 'app_eleve_show', methods: ['GET'])]
     public function showEleve(Eleve $eleve): Response
     {
@@ -61,17 +63,16 @@ final class UsersController extends AbstractController
         ]);
     }
 
-    // Modification d'un élève existant
     #[Route('/Eleve/{id}/edit', name: 'app_eleve_edit', methods: ['GET', 'POST'])]
     public function editEleve(Request $request, Eleve $eleve, EntityManagerInterface $entityManager): Response
     {
-        $form = $this->createForm(EleveType::class, $eleve);
+        $form = $this->createEleveForm($eleve);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_users', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToUsersIndex();
         }
 
         return $this->render('eleve/edit.html.twig', [
@@ -80,35 +81,39 @@ final class UsersController extends AbstractController
         ]);
     }
 
-
-    // Suppression d'un élève
     #[Route('/Eleve/{id}', name: 'app_eleve_delete', methods: ['POST'])]
-    public function deleteEleve(Request $request, Eleve $eleve, EntityManagerInterface $entityManager): Response
+    public function deleteEleve(Request $request, Eleve $eleve, EntityManagerInterface $entityManager, StageRepository $stageRepository): Response
     {
         if ($this->isCsrfTokenValid('delete'.$eleve->getId(), $request->getPayload()->getString('_token'))) {
-            $entityManager->remove($eleve);
-            $entityManager->flush();
+            if ($stageRepository->findOneBy(['eleveStage' => $eleve]) !== null) {
+                $this->addFlash('warning', 'Impossible de supprimer cet élève : un stage correspond déjà à cet élève.');
+
+                return $this->redirectToUsersIndex();
+            }
+
+            try {
+                $entityManager->remove($eleve);
+                $entityManager->flush();
+            } catch (ForeignKeyConstraintViolationException) {
+                $this->addFlash('warning', 'Impossible de supprimer cet élève : un stage correspond déjà à cet élève.');
+            }
         }
 
-        return $this->redirectToRoute('app_users', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToUsersIndex();
     }
 
-
-
-    // CRUD pour les utilisateurs
-    // Création d'un nouvel utilisateur
     #[Route('/newUtilisateur', name: 'app_utilisateur_new', methods: ['GET', 'POST'])]
     public function newUtilisateur(Request $request, EntityManagerInterface $entityManager): Response
     {
         $utilisateur = new Utilisateur();
-        $form = $this->createForm(UtilisateurType::class, $utilisateur);
+        $form = $this->createUtilisateurForm($utilisateur, true);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->persist($utilisateur);
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_users', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToUsersIndex();
         }
 
         return $this->render('utilisateur/new.html.twig', [
@@ -117,7 +122,6 @@ final class UsersController extends AbstractController
         ]);
     }
 
-    // Affichage des détails d'un utilisateur
     #[Route('/Utilisateur/{id}', name: 'app_utilisateur_show', methods: ['GET'])]
     public function showUtilisateur(Utilisateur $utilisateur): Response
     {
@@ -126,19 +130,16 @@ final class UsersController extends AbstractController
         ]);
     }
 
-    // Modification d'un utilisateur existant
     #[Route('/Utilisateur/{id}/edit', name: 'app_utilisateur_edit', methods: ['GET', 'POST'])]
     public function editUtilisateur(Request $request, Utilisateur $utilisateur, EntityManagerInterface $entityManager): Response
     {
-        $form = $this->createForm(UtilisateurType::class, $utilisateur, [
-            'show_password' => false,
-        ]);
+        $form = $this->createUtilisateurForm($utilisateur, false);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->flush();
 
-            return $this->redirectToRoute('app_users', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToUsersIndex();
         }
 
         return $this->render('utilisateur/edit.html.twig', [
@@ -147,7 +148,6 @@ final class UsersController extends AbstractController
         ]);
     }
 
-    // Suppression d'un utilisateur
     #[Route('/Utilisateur/{id}', name: 'app_utilisateur_delete', methods: ['POST'])]
     public function deleteUtilisateur(Request $request, Utilisateur $utilisateur, EntityManagerInterface $entityManager): Response
     {
@@ -156,12 +156,29 @@ final class UsersController extends AbstractController
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('app_users', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToUsersIndex();
     }
 
     #[Route('/access-denied', name: 'app_access_denied')]
     public function accessDenied(): Response
     {
         return $this->render('security/access_denied.html.twig');
+    }
+
+    private function redirectToUsersIndex(): Response
+    {
+        return $this->redirectToRoute('app_users', [], Response::HTTP_SEE_OTHER);
+    }
+
+    private function createEleveForm(Eleve $eleve)
+    {
+        return $this->createForm(EleveType::class, $eleve);
+    }
+
+    private function createUtilisateurForm(Utilisateur $utilisateur, bool $showPassword)
+    {
+        return $this->createForm(UtilisateurType::class, $utilisateur, [
+            'show_password' => $showPassword,
+        ]);
     }
 }
